refactor(cell): split use_ability.rs (902 lines) into directory module
Extract test module into sibling file:

- mod.rs: ability dispatch + handler logic (538 lines)
- tests.rs: unit tests (364 lines)
